backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
reportlab
python-multipart
fastapi-cors
psycopg2-binary
pydantic[email]
pydantic-settings # <-- AÑADIMOS LA LIBRERÍA QUE FALTABA
python-dateutil
bcrypt==3.2.2
cryptography
num2words
qrcode[pil]